requirements.txt
Flask
requests
torch
transformers
sentence-transformers
faiss-cpu==1.10.0
scikit-learn
matplotlib
gunicorn
